refactor: move what the hub shares with admin into common
After the tier flip admin and the embedding hub are both feature tier,
and a feature module may not import another, so anything both read has
to surface into a shared module.

base-path.ts is that: six admin files, three enterprise files and two
hub files read it, so it moves to common along with its spec. The
public-links table moves for the same reason -- admin's Public sharing
page and EmbeddedResources both render it, and EmbeddedResources itself
moves to embedding/settings since only the hub's Security page uses it.

EmbeddingHubPermissionsBasePath hardcodes the hub's path and is read by
the hub alone, so it moves into the hub rather than into common. Every
other file here is a one to three line import repoint.
