+++ frontend/src/metabase/admin/permissions/pages/CollectionPermissionsPage/CollectionPermissionsPage.tsx
@@ -3,6 +3,7 @@ import { t } from "ttag";
 
 import { CollectionPermissionsHelp } from "metabase/admin/permissions/components/CollectionPermissionsHelp";
 import { useListCollectionsTreeQuery } from "metabase/api";
+import { getPermissionsBasePath } from "metabase/common/components/PermissionsBasePath/base-path";
 import { useDispatch, useSelector } from "metabase/redux";
 import { useNavigate, useParams } from "metabase/router";
 import type { CollectionId } from "metabase-types/api";
@@ -28,7 +29,6 @@ import {
 } from "../../selectors/collection-permissions";
 import type { PermissionEditorEntity } from "../../types";
 import { assertNumericId } from "../../types";
-import { getPermissionsBasePath } from "../../utils/base-path";
 
 export function CollectionPermissionsPage() {
   const dispatch = useDispatch();
