+++ frontend/src/metabase/admin/permissions/utils/is-embedding-hub.ts
@@ -1,4 +1,4 @@
-import { getPermissionsBasePath } from "metabase/admin/permissions/utils/base-path";
+import { getPermissionsBasePath } from "metabase/common/components/PermissionsBasePath/base-path";
 import * as Urls from "metabase/urls";
 
 /** Used to pick admin's vs. the hub's color scheme, based on where the permissions editor is mounted. */
