+++ frontend/src/metabase/admin/settings/components/widgets/PublicLinksListing/PublicDocumentResources.tsx
@@ -1,10 +1,10 @@
 import { t } from "ttag";
 
-import { PublicLinksListing } from "metabase/admin/settings/components/widgets/PublicLinksListing/index";
 import {
   useDeleteDocumentPublicLinkMutation,
   useListPublicDocumentsQuery,
 } from "metabase/api";
+import { PublicLinksListing } from "metabase/common/components/PublicLinksListing/PublicLinksListing";
 import * as Urls from "metabase/urls";
 import type { Document } from "metabase-types/api";
