+++ frontend/src/metabase/admin/settings/components/widgets/PublicLinksListing/PublicResources.tsx
@@ -8,6 +8,7 @@ import {
   useListPublicCardsQuery,
   useListPublicDashboardsQuery,
 } from "metabase/api";
+import { PublicLinksListing } from "metabase/common/components/PublicLinksListing/PublicLinksListing";
 import { useSetting } from "metabase/settings";
 import * as Urls from "metabase/urls";
 import type {
@@ -15,8 +16,6 @@ import type {
   GetPublicCard,
   GetPublicDashboard,
 } from "metabase-types/api";
-
-import { PublicLinksListing } from "./PublicLinksListing";
 
 export const PublicLinksDashboardListing = () => {
   const query = useListPublicDashboardsQuery();
