+++ frontend/src/metabase/admin/settings/components/widgets/PublicLinksListing/index.ts
@@ -1,4 +1,2 @@
-export * from "./PublicLinksListing";
-export * from "./EmbeddedResources";
 export * from "./PublicResources";
 export * from "./PublicDocumentResources";
